+++ server/controllers/auth.controller.js
@@ -1,7 +1,7 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 
-const User = require('../models/user');
+const User = require('../models/user.model');
 
 const { JWT_SECRET } = process.env;
