+++ server/middleware/auth.jwt.js
@@ -1,5 +1,6 @@
 require('dotenv').config();
 const jwt = require('jsonwebtoken');
+// const User = require('../models/user');
 
 const verifyToken = (req, res, next) => {
   const bearerHeader = req.headers.authorization;
@@ -17,5 +18,22 @@ const verifyToken = (req, res, next) => {
     return next();
   });
 };
+
+/* const isAdmin = (req, res, next) => {
+  User.findById(req.userId).then((user) => {
+    user. ().then((roles) => {
+      for (let i = 0; i < roles.length; i++) {
+        if (roles[i].name === 'admin') {
+          next();
+          return;
+        }
+      }
+
+      res.status(403).send({
+        message: 'Require Admin Role!',
+      });
+    });
+  });
+}; */
 
 module.exports = verifyToken;
