+++ server/routes/auth.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 
-const authController = require('../controllers/auth');
+const authController = require('../controllers/auth.controller');
 const { validate } = require('../middleware/validate');
 const { checkEmailExists } = require('../middleware/verifySignUp');
 const {
@@ -12,13 +12,13 @@ const app = express();
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-app.use((req, res, next) => {
+/* app.use((req, res, next) => {
   res.header(
     'Access-Control-Allow-Headers',
     'x-access-token, Origin, Content-Type, Accept',
   );
   next();
-});
+}); */
 
 // check('what in req', 'message if error').typeOfCheck()
 app.post('/signin', validate(validateUserLogin), authController.signIn);
