--- server/routes/users.js
@@ -1,17 +1,0 @@
-const express = require('express');
-const userController = require('../controllers/user');
-const notFound = require('../controllers/notFoundController');
-const verifyToken = require('../middleware/auth.verify.token');
-
-const router = express();
-
-router.use(express.json());
-router.use(express.urlencoded({ extended: false }));
-
-router.get('/', verifyToken, userController.index);
-router.get('/find/:email', verifyToken, userController.detail);
-router.get('/*', notFound);
-
-router.post('/find', verifyToken, userController.detail);
-
-module.exports = router;
